verbs: Fix possible port loop overflow
[ Upstream commit 348988dc970b36a6ee93271f1ca0f122731967df ]

When going over all the ports of an IB device and a uint_t8 variable is
used a possible overflow can happen if the IB device has 255 ports.
Once overflow occurs an invalid port (i.e. 0) is used.
Fix by switching to uint32_t.

Fixes: 24b7e9ddd4e4 ("Merge Dotan Barak's vstat tool into ibv_devinfo")

+++ libibverbs/examples/devinfo.c
@@ -162,7 +162,7 @@ static const char *vl_str(uint8_t vl_num)
 	}
 }
 
-static int print_all_port_gids(struct ibv_context *ctx, uint8_t port_num, int tbl_len)
+static int print_all_port_gids(struct ibv_context *ctx, uint32_t port_num, int tbl_len)
 {
 	union ibv_gid gid;
 	int rc = 0;
@@ -171,7 +171,7 @@ static int print_all_port_gids(struct ibv_context *ctx, uint8_t port_num, int tb
 	for (i = 0; i < tbl_len; i++) {
 		rc = ibv_query_gid(ctx, port_num, i, &gid);
 		if (rc) {
-			fprintf(stderr, "Failed to query gid to port %d, index %d\n",
+			fprintf(stderr, "Failed to query gid to port %u, index %d\n",
 			       port_num, i);
 			return rc;
 		}
@@ -461,7 +461,7 @@ static int print_hca_cap(struct ibv_device *ib_dev, uint8_t ib_port)
 	struct ibv_device_attr_ex device_attr;
 	struct ibv_port_attr port_attr;
 	int rc = 0;
-	uint8_t port;
+	uint32_t port;
 	char buf[256];
 
 	ctx = ibv_open_device(ib_dev);
@@ -580,7 +580,7 @@ static int print_hca_cap(struct ibv_device *ib_dev, uint8_t ib_port)
 			fprintf(stderr, "Failed to query port %u props\n", port);
 			goto cleanup;
 		}
-		printf("\t\tport:\t%d\n", port);
+		printf("\t\tport:\t%u\n", port);
 		printf("\t\t\tstate:\t\t\t%s (%d)\n",
 		       port_state_str(port_attr.state), port_attr.state);
 		printf("\t\t\tmax_mtu:\t\t%s (%d)\n",
